Removed Batch#register moethod and moved all batch methods to the Batch class.
