+++ src/model/delta/insertdelta.js
@@ -10,13 +10,6 @@
 import Delta from './delta';
 import RemoveDelta from './removedelta';
 import DeltaFactory from './deltafactory';
-import InsertOperation from '../operation/insertoperation';
-import { register } from '../batch';
-import { normalizeNodes } from './../writer';
-
-import DocumentFragment from '../documentfragment';
-import Range from '../../model/range.js';
-import Position from '../../model/position.js';
 
 /**
  * To provide specific OT behavior and better collisions solving, the {@link module:engine/model/batch~Batch#insert Batch#insert} method
@@ -77,48 +70,5 @@ export default class InsertDelta extends Delta {
 		return 'engine.model.delta.InsertDelta';
 	}
 }
-
-/**
- * Inserts a node or nodes at the given position.
- *
- * When inserted element is a {@link module:engine/model/documentfragment~DocumentFragment} and has markers its markers will
- * be set to {@link module:engine/model/document~Document#markers}.
- *
- * @chainable
- * @method module:engine/model/batch~Batch#insert
- * @param {module:engine/model/position~Position} position Position of insertion.
- * @param {module:engine/model/node~NodeSet} nodes The list of nodes to be inserted.
- */
-register( 'insert', function( position, nodes ) {
-	const normalizedNodes = normalizeNodes( nodes );
-
-	// If nothing is inserted do not create delta and operation.
-	if ( normalizedNodes.length === 0 ) {
-		return this;
-	}
-
-	const delta = new InsertDelta();
-	const insert = new InsertOperation( position, normalizedNodes, this.document.version );
-
-	this.addDelta( delta );
-	delta.addOperation( insert );
-	this.document.applyOperation( insert );
-
-	// When element is a DocumentFragment we need to move its markers to Document#markers.
-	if ( nodes instanceof DocumentFragment ) {
-		for ( const [ markerName, markerRange ] of nodes.markers ) {
-			// We need to migrate marker range from DocumentFragment to Document.
-			const rangeRootPosition = Position.createAt( markerRange.root );
-			const range = new Range(
-				markerRange.start._getCombined( rangeRootPosition, position ),
-				markerRange.end._getCombined( rangeRootPosition, position )
-			);
-
-			this.setMarker( markerName, range );
-		}
-	}
-
-	return this;
-} );
 
 DeltaFactory.register( InsertDelta );
