+++ src/model/delta/movedelta.js
@@ -9,11 +9,6 @@
 
 import Delta from './delta';
 import DeltaFactory from './deltafactory';
-import { register } from '../batch';
-import MoveOperation from '../operation/moveoperation';
-import Position from '../position';
-import Range from '../range';
-import CKEditorError from '@ckeditor/ckeditor5-utils/src/ckeditorerror';
 
 /**
  * To provide specific OT behavior and better collisions solving, {@link module:engine/model/batch~Batch#move} method
@@ -85,41 +80,5 @@ export default class MoveDelta extends Delta {
 		return 'engine.model.delta.MoveDelta';
 	}
 }
-
-function addMoveOperation( batch, delta, sourcePosition, howMany, targetPosition ) {
-	const operation = new MoveOperation( sourcePosition, howMany, targetPosition, batch.document.version );
-	delta.addOperation( operation );
-	batch.document.applyOperation( operation );
-}
-
-/**
- * Moves given {@link module:engine/model/item~Item model item} or given range to target position.
- *
- * @chainable
- * @method module:engine/model/batch~Batch#move
- * @param {module:engine/model/item~Item|module:engine/model/range~Range} itemOrRange Model item or range of nodes to move.
- * @param {module:engine/model/position~Position} targetPosition Position where moved nodes will be inserted.
- */
-register( 'move', function( itemOrRange, targetPosition ) {
-	const delta = new MoveDelta();
-	this.addDelta( delta );
-
-	if ( itemOrRange instanceof Range ) {
-		if ( !itemOrRange.isFlat ) {
-			/**
-			 * Range to move is not flat.
-			 *
-			 * @error batch-move-range-not-flat
-			 */
-			throw new CKEditorError( 'batch-move-range-not-flat: Range to move is not flat.' );
-		}
-
-		addMoveOperation( this, delta, itemOrRange.start, itemOrRange.end.offset - itemOrRange.start.offset, targetPosition );
-	} else {
-		addMoveOperation( this, delta, Position.createBefore( itemOrRange ), 1, targetPosition );
-	}
-
-	return this;
-} );
 
 DeltaFactory.register( MoveDelta );
