+++ src/model/delta/unwrapdelta.js
@@ -10,11 +10,6 @@
 import Delta from './delta';
 import DeltaFactory from './deltafactory';
 import WrapDelta from './wrapdelta';
-import { register } from '../batch';
-import Position from '../position';
-import RemoveOperation from '../operation/removeoperation';
-import MoveOperation from '../operation/moveoperation';
-import CKEditorError from '@ckeditor/ckeditor5-utils/src/ckeditorerror';
 
 /**
  * To provide specific OT behavior and better collisions solving, {@link module:engine/model/batch~Batch#merge} method
@@ -63,51 +58,5 @@ export default class UnwrapDelta extends Delta {
 		return 'engine.model.delta.UnwrapDelta';
 	}
 }
-
-/**
- * Unwraps children of the given element – all its children are moved before it and then the element is removed.
- * Throws error if you try to unwrap an element which does not have a parent.
- *
- * @chainable
- * @method module:engine/model/batch~Batch#unwrap
- * @param {module:engine/model/element~Element} position Element to unwrap.
- */
-register( 'unwrap', function( element ) {
-	if ( element.parent === null ) {
-		/**
-		 * Trying to unwrap an element which has no parent.
-		 *
-		 * @error batch-unwrap-element-no-parent
-		 */
-		throw new CKEditorError( 'batch-unwrap-element-no-parent: Trying to unwrap an element which has no parent.' );
-	}
-
-	const delta = new UnwrapDelta();
-	this.addDelta( delta );
-
-	const sourcePosition = Position.createFromParentAndOffset( element, 0 );
-
-	const move = new MoveOperation(
-		sourcePosition,
-		element.maxOffset,
-		Position.createBefore( element ),
-		this.document.version
-	);
-
-	move.isSticky = true;
-	delta.addOperation( move );
-	this.document.applyOperation( move );
-
-	// Computing new position because we moved some nodes before `element`.
-	// If we would cache `Position.createBefore( element )` we remove wrong node.
-	const graveyard = this.document.graveyard;
-	const gyPosition = new Position( graveyard, [ 0 ] );
-
-	const remove = new RemoveOperation( Position.createBefore( element ), 1, gyPosition, this.document.version );
-	delta.addOperation( remove );
-	this.document.applyOperation( remove );
-
-	return this;
-} );
 
 DeltaFactory.register( UnwrapDelta );
