+++ tests/model/delta/wrapdelta.js
@@ -5,10 +5,7 @@
 
 import Document from '../../../src/model/document';
 import Position from '../../../src/model/position';
-import Range from '../../../src/model/range';
 import Element from '../../../src/model/element';
-import Text from '../../../src/model/text';
-import CKEditorError from '@ckeditor/ckeditor5-utils/src/ckeditorerror';
 
 import WrapDelta from '../../../src/model/delta/wrapdelta';
 import UnwrapDelta from '../../../src/model/delta/unwrapdelta';
@@ -16,86 +13,6 @@ import UnwrapDelta from '../../../src/model/delta/unwrapdelta';
 import InsertOperation from '../../../src/model/operation/insertoperation';
 import MoveOperation from '../../../src/model/operation/moveoperation';
 import RemoveOperation from '../../../src/model/operation/removeoperation';
-
-describe( 'Batch', () => {
-	let doc, root, range;
-
-	beforeEach( () => {
-		doc = new Document();
-		root = doc.createRoot();
-
-		root.insertChildren( 0, new Text( 'foobar' ) );
-
-		range = new Range( new Position( root, [ 2 ] ), new Position( root, [ 4 ] ) );
-	} );
-
-	describe( 'wrap', () => {
-		it( 'should wrap flat range with given element', () => {
-			const p = new Element( 'p' );
-			doc.batch().wrap( range, p );
-
-			expect( root.maxOffset ).to.equal( 5 );
-			expect( root.getChild( 0 ).data ).to.equal( 'fo' );
-			expect( root.getChild( 1 ) ).to.equal( p );
-			expect( p.getChild( 0 ).data ).to.equal( 'ob' );
-			expect( root.getChild( 2 ).data ).to.equal( 'ar' );
-		} );
-
-		it( 'should wrap flat range with an element of given name', () => {
-			doc.batch().wrap( range, 'p' );
-
-			expect( root.maxOffset ).to.equal( 5 );
-			expect( root.getChild( 0 ).data ).to.equal( 'fo' );
-			expect( root.getChild( 1 ).name ).to.equal( 'p' );
-			expect( root.getChild( 1 ).getChild( 0 ).data ).to.equal( 'ob' );
-			expect( root.getChild( 2 ).data ).to.equal( 'ar' );
-		} );
-
-		it( 'should throw if range to wrap is not flat', () => {
-			root.insertChildren( 1, [ new Element( 'p', [], new Text( 'xyz' ) ) ] );
-			const notFlatRange = new Range( new Position( root, [ 3 ] ), new Position( root, [ 6, 2 ] ) );
-
-			expect( () => {
-				doc.batch().wrap( notFlatRange, 'p' );
-			} ).to.throw( CKEditorError, /^batch-wrap-range-not-flat/ );
-		} );
-
-		it( 'should throw if element to wrap with has children', () => {
-			const p = new Element( 'p', [], new Text( 'a' ) );
-
-			expect( () => {
-				doc.batch().wrap( range, p );
-			} ).to.throw( CKEditorError, /^batch-wrap-element-not-empty/ );
-		} );
-
-		it( 'should throw if element to wrap with has children', () => {
-			const p = new Element( 'p' );
-			root.insertChildren( 0, p );
-
-			expect( () => {
-				doc.batch().wrap( range, p );
-			} ).to.throw( CKEditorError, /^batch-wrap-element-attached/ );
-		} );
-
-		it( 'should be chainable', () => {
-			const batch = doc.batch();
-
-			const chain = batch.wrap( range, 'p' );
-			expect( chain ).to.equal( batch );
-		} );
-
-		it( 'should add delta to batch and operation to delta before applying operation', () => {
-			sinon.spy( doc, 'applyOperation' );
-			const batch = doc.batch().wrap( range, 'p' );
-
-			const correctDeltaMatcher = sinon.match( operation => {
-				return operation.delta && operation.delta.batch && operation.delta.batch == batch;
-			} );
-
-			expect( doc.applyOperation.calledWith( correctDeltaMatcher ) ).to.be.true;
-		} );
-	} );
-} );
 
 describe( 'WrapDelta', () => {
 	let wrapDelta, doc, root;
